% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeannotate.R
\name{codeannotate}
\alias{codeannotate}
\title{Code Annotate}
\usage{
codeannotate()
}
\description{
Add code annotation markup to a code chunk in a quarto document.
}
\details{
After installing this RStudio addin, when lines of code are selected,
the markup for \href{https://quarto.org/docs/authoring/code-annotation.html}{quarto code annotations}
will be added.
}
\seealso{
\code{\link[=remove_codeannotate]{remove_codeannotate()}}
}
